+++ src/components/ExampleComponent/tests/index.js
@@ -5,7 +5,7 @@ import { mountToJson } from 'enzyme-to-json';
 import Input from '../../Input';
 import Button from '../../Button';
 
-describe('components/ExampleComponent', () => {
+describe('components/ExampleComponent', function() {
   let component;
   let fnMock = jest.fn(event => { return event; });
 
@@ -43,39 +43,39 @@ describe('components/ExampleComponent', () => {
     );
   });
 
-  it('should update "aStringValue" prop', () => {
+  it('should update "aStringValue" prop', function() {
     expect(component.find('[data-test-section="aString"]').text()).toBe('aStringValue (with updates)');
   });
 
-  it('should update "aNumber" prop', () => {
+  it('should update "aNumber" prop', function() {
     expect(component.find('[data-test-section="aNumber"]').text()).toBe('6');
   });
 
-  it('should update "aBoolean" prop', () => {
+  it('should update "aBoolean" prop', function() {
     expect(component.find('[data-test-section="aBoolean"]').text()).toBe('false');
   });
 
-  it('should update "anArray" prop', () => {
+  it('should update "anArray" prop', function() {
     expect(component.find('[data-test-section="anArray"]').text()).toBe('[1,2,3,4,5,6]');
   });
 
-  it('should update "anObject" prop', () => {
+  it('should update "anObject" prop', function() {
     expect(component.find('[data-test-section="anObject"]').text()).toBe('{\"initialValue\":\"INITIAL_VALUE\",\"newKey\":true}');
   });
 
-  it('should update ExampleComponent state and return name to children when a user types their name in an input', () => {
+  it('should update ExampleComponent state and return name to children when a user types their name in an input', function() {
     expect(component.state()).toMatchObject({ name: ''});
     component.find('[data-test-section="nameInput"]').simulate('input', { target: { value: 'Johnny' } });
     expect(component.state()).toMatchObject({ name: 'Johnny'});
     expect(component.find('[data-test-section="inputNameValue"]').text()).toBe('Johnny');
   });
 
-  it('should call aFunction with a string argument of "updated" when the Button is pressed', () => {
+  it('should call aFunction with a string argument of "updated" when the Button is pressed', function() {
     component.find('[data-test-section="pressMe"]').simulate('click');
     expect(fnMock.mock.calls[0][0]).toBe('updated');
   });
 
-  it('should match the Jest Snapshot', () => {
+  it('should match the Jest Snapshot', function() {
     expect(mountToJson(component)).toMatchSnapshot();
   });
 });